// src/types/env.ts
export interface IENV {
    VITE_IAM_CLIENT_SECRET: string;
    VITE_BNP_API_KEY: string;

    VITE_IAM_BASE_URL: string;
    VITE_DEV_API_SOURCE: string;
    VITE_BNP_BASE_URL: string;
    VITE_CRM_BASE_URL: string;

    VITE_IAM_TOKEN_ENDPOINT: string;
    VITE_IAM_LOGOUT_ENDPOINT: string;
    VITE_IAM_RESET_PASSWORD_WITH_CODE_ENDPOINT: string;
    VITE_IAM_VALIDATE_CODE_FOR_PASSWORD_RESET_ENDPOINT: string;
    VITE_IAM_CODE_FOR_PASSWORD_RESET_ENDPOINT: string;
    VITE_BNP_INVOICE_ENDPOINT: string;
}